requirements.txt
# Core dependencies
langchain==0.0.298
qdrant-client==1.6.4
sentence-transformers==2.2.2
huggingface-hub==0.16.4
transformers==4.30.2
torch==2.0.1
python-dotenv==1.0.0

# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2

# Utilities
numpy==1.24.3
requests==2.31.0
python-multipart==0.0.6
typing-extensions==4.8.0
colorama==0.4.6

# Testing and Development

# Development dependencies
pytest==7.4.3
black==23.10.1
isort==5.12.0
flake8==6.1.0
mypy==1.6.1

# Logging and monitoring
python-json-logger==2.0.7
